rcm/tag-darwin/Alfred.alfredpreferences/workflows/lib/requirements.txt
Babel==2.9.1
iso8601==0.1.13
pyyaml==5.4.1
requests-cache==0.5.2
requests==2.31.0
richxerox==1.0.1
tacky==0.0.4
url-normalize==1.4.3
yq==2.12.0
